% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_BuildupEffect.r
\name{BuildupEffect}
\alias{BuildupEffect}
\title{Build Up Effect Calculator}
\usage{
BuildupEffect(FUELTYPE, BUI)
}
\arguments{
\item{FUELTYPE}{The Fire Behaviour Prediction FuelType}

\item{BUI}{The Buildup Index value}
}
\value{
BE: Build up effect
}
\description{
Computes the Buildup Effect on Fire Spread Rate. All variables names are 
laid out in the same manner as Forestry Canada Fire Danger Group (FCFDG) 
(1992).
}
\references{
\url{https://cfs.nrcan.gc.ca/publications/download-pdf/10068} 
Development and Structure of the Canadian Forest Fire Behavior Prediction 
System." Technical Report ST-X-3, Forestry Canada, Ottawa, Ontario.
}
