% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildupIndex.r
\name{BuildupIndex}
\alias{BuildupIndex}
\title{Build Up Index Calculator}
\usage{
BuildupIndex(dmc, dc)
}
\arguments{
\item{dmc}{Duff Moisture Code}

\item{dc}{Drought Code}
}
\value{
A single Build Up Index value
}
\description{
Buildup Index Calculation. All code is based on a C code library that was 
written by Canadian Forest Service Employees, which was originally based on 
the Fortran code listed in the reference below. All equations in this code 
refer to that document.
Equations and FORTRAN program for the Canadian Forest Fire Weather Index 
System. 1985. Van Wagner, C.E.; Pickett, T.L. Canadian Forestry Service, 
Petawawa National Forestry Institute, Chalk River, Ontario. Forestry 
Technical Report 33. 18 p.
}
\references{
\url{http://cfs.nrcan.gc.ca/pubwarehouse/pdfs/19927.pdf} 
Development and structure of the Canadian Forest Fire Weather Index System. 
1987. Van Wagner, C.E. Canadian Forestry Service, Headquarters, Ottawa. 
Forestry Technical Report 35. 35 p.
}
