% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CriticalSurfaceIntensity.r
\name{CriticalSurfaceIntensity}
\alias{CriticalSurfaceIntensity}
\title{Critical Surface Intensity Calculator}
\usage{
CriticalSurfaceIntensity(FMC, CBH)
}
\arguments{
\item{FMC}{Foliar Moisture Content}

\item{CBH}{Crown Base Height}
}
\value{
CSI
}
\description{
Calculate Critical surface intensity (CSI).
}
\details{
All variables names are laid out in the same manner as Forestry Canada Fire 
Danger Group (FCFDG) (1992). Development and Structure of the Canadian Forest
 Fire Behavior Prediction System." Technical Report ST-X-3, Forestry Canada,
  Ottawa, Ontario.
}
\references{
\url{https://cfs.nrcan.gc.ca/publications/download-pdf/10068} 
Development and Structure of the Canadian Forest Fire Behavior Prediction 
System." Technical Report ST-X-3, Forestry Canada, Ottawa, Ontario.
}
