% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_CriticalSurfaceRateOfSpread.r
\name{CriticalSurfaceRateOfSpread}
\alias{CriticalSurfaceRateOfSpread}
\title{Critical Surface Rate of Spread Calculator}
\usage{
CriticalSurfaceRateOfSpread(CSI, SFC)
}
\arguments{
\item{CSI}{Critical Surface Intensity}

\item{SFC}{Surface Fuel Consumption}
}
\value{
RSO
}
\description{
Calculate Critical Surface fire rate of spread (RSO). The value of each of these equations can be returned to the calling 
function without unecessary additional calculations.
}
\details{
All variables names are laid out in the same manner as Forestry Canada Fire 
Danger Group (FCFDG) (1992). Development and Structure of the Canadian Forest
 Fire Behavior Prediction System." Technical Report ST-X-3, Forestry Canada,
  Ottawa, Ontario.
}
\references{
\url{https://cfs.nrcan.gc.ca/publications/download-pdf/10068} 
Development and Structure of the Canadian Forest Fire Behavior Prediction 
System." Technical Report ST-X-3, Forestry Canada, Ottawa, Ontario.
}
