% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrownFractionBurned.r
\name{CrownFractionBurned}
\alias{CrownFractionBurned}
\title{Crown Fraction Burned Calculator}
\usage{
CrownFractionBurned(FUELTYPE, FMC, SFC, ROS, CBH, option = "CFB")
}
\arguments{
\item{FUELTYPE}{The Fire Behaviour Prediction FuelType}

\item{FMC}{Foliar Moisture Content}

\item{SFC}{Surface Fuel Consumption}

\item{ROS}{Rate of Spread}

\item{CBH}{Crown Base Height}

\item{option}{Which variable to calculate(ROS, CFB, RSC, or RSI)}
}
\value{
CFB, CSI, RSO depending on which option was selected. _Default:_ "CFB"
}
\description{
Calculate Calculate Crown Fraction Burned. To calculate CFB, we also need to 
calculate Critical surface intensity (CSI), and Surface fire rate of spread 
(RSO). The value of each of these equations can be returned to the calling 
function without unecessary additional calculations.
}
\details{
All variables names are laid out in the same manner as Forestry Canada Fire 
Danger Group (FCFDG) (1992). Development and Structure of the Canadian Forest
 Fire Behavior Prediction System." Technical Report ST-X-3, Forestry Canada,
  Ottawa, Ontario.
}
\references{
\url{https://cfs.nrcan.gc.ca/publications/download-pdf/10068} 
Development and Structure of the Canadian Forest Fire Behavior Prediction 
System." Technical Report ST-X-3, Forestry Canada, Ottawa, Ontario.
}
