% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrownRateOfSpreadC6.R
\name{CrownRateOfSpreadC6}
\alias{CrownRateOfSpreadC6}
\title{C-6 Crown Fire Spread Calculator}
\usage{
CrownRateOfSpreadC6(ISI, FMC)
}
\arguments{
\item{ISI}{Initial Spread Index}
}
\value{
RSC
}
\description{
Calculate crown rate of spread (RSC).
}
\details{
All variables names are laid out in the same manner as Forestry Canada Fire 
Danger Group (FCFDG) (1992). Development and Structure of the Canadian Forest
 Fire Behavior Prediction System." Technical Report ST-X-3, Forestry Canada, 
 Ottawa, Ontario.
}
\references{
\url{https://cfs.nrcan.gc.ca/publications/download-pdf/10068} 
Development and Structure of the Canadian Forest Fire Behavior Prediction 
System." Technical Report ST-X-3, Forestry Canada, Ottawa, Ontario.
}
