% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Direction.r
\name{Direction}
\alias{Direction}
\title{Direction definer}
\usage{
Direction(bearingT1T2, bearingT1T3, ThetaAdeg)
}
\arguments{
\item{bearingT1T2}{Bearing between T1 and T2}

\item{bearingT1T3}{Bearing between T1 and T3}

\item{ThetaAdeg}{Direction}
}
\value{
DIR - a direction in degrees
}
\description{
New DIRECTION function to determine clockwise or counter-clockwise 
"interpretation"
}
