% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceAtTime.r
\name{DistanceAtTime}
\alias{DistanceAtTime}
\title{Distance at time t calculator}
\usage{
DistanceAtTime(FUELTYPE, ROSeq, HR, CFB)
}
\arguments{
\item{FUELTYPE}{The Fire Behaviour Prediction FuelType}

\item{ROSeq}{The predicted equilibrium rate of spread (m/min)}

\item{HR}{The elapsed time (min)}

\item{CFB}{Crown Fraction Burned}
}
\value{
DISTt - Head fire spread distance at time t
}
\description{
Calculate the Head fire spread distance at time t. In the documentation this 
variable is just "D".
}
\details{
All variables names are laid out in the same manner as Forestry Canada Fire 
Danger Group (FCFDG) (1992). Development and Structure of the  Canadian 
Forest Fire Behavior Prediction System." Technical Report ST-X-3, 
Forestry Canada, Ottawa, Ontario.
}
