% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dmc.r
\name{Dmc}
\alias{Dmc}
\title{Duff Moisture Code Calculator}
\usage{
Dmc(dmc_yda, temp, rh, prec, lat, mon, lat.adjust = TRUE)
}
\arguments{
\item{dmc_yda}{The Duff Moisture Code from previous iteration}

\item{temp}{Temperature (centigrade)}

\item{rh}{Relative Humidity (%)}

\item{prec}{Precipitation(mm)}

\item{lat}{Latitude (decimal degrees)}

\item{mon}{Month (1-12)}

\item{lat.adjust}{Latitude adjustment (TRUE, FALSE, default=TRUE)}
}
\value{
A single drought moisture code value
}
\description{
Duff Moisture Code Calculation. All code is based on a C code library that 
was written by Canadian Forest Service Employees, which was originally based
on the Fortran code listed in the reference below. All equations in this code
 refer to that document.
}
\details{
Equations and FORTRAN program for the Canadian Forest Fire Weather Index 
System. 1985. Van Wagner, C.E.; Pickett, T.L. Canadian Forestry Service, 
Petawawa National Forestry Institute, Chalk River, Ontario. Forestry 
Technical Report 33. 18 p.

Additional reference on FWI system

Development and structure of the Canadian Forest Fire Weather Index System. 
1987. Van Wagner, C.E. Canadian Forestry Service, Headquarters, Ottawa. 
Forestry Technical Report 35. 35 p.
}
\references{
\url{http://cfs.nrcan.gc.ca/pubwarehouse/pdfs/19927.pdf} 
Development and structure of the Canadian Forest Fire Weather Index System. 
1987. Van Wagner, C.E. Canadian Forestry Service, Headquarters, Ottawa. 
Forestry Technical Report 35. 35 p.
}
