% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BROScalc.r
\name{.BROScalc}
\alias{.BROScalc}
\title{Back Fire Rate of Spread Calculator}
\usage{
.BROScalc(FUELTYPE, FFMC, BUI, WSV, FMC, SFC, PC, PDF, CC, CBH)
}
\arguments{
\item{FUELTYPE}{The Fire Behaviour Prediction FuelType}

\item{FFMC}{Fine Fuel Moisture Code}

\item{BUI}{Buildup Index}

\item{WSV}{Wind Speed Vector}

\item{FMC}{Foliar Moisture Content}

\item{SFC}{Surface Fuel Consumption}

\item{PC}{Percent Conifer}

\item{PDF}{Percent Dead Balsam Fir}

\item{CC}{Degree of Curing (just "C" in FCFDG 1992)}

\item{CBH}{Crown Base Height}
}
\value{
BROS: Back Fire Rate of Spread
}
\description{
Calculate the Back Fire Spread Rate. All variables names are laid out in the 
same manner as Forestry Canada Fire Danger Group (FCFDG) (1992).
}
\references{
\url{https://cfs.nrcan.gc.ca/publications/download-pdf/10068} 
Development and Structure of the Canadian Forest Fire Behavior Prediction 
System." Technical Report ST-X-3, Forestry Canada, Ottawa, Ontario.
}
