% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwiCalc.r
\name{.fwiCalc}
\alias{.fwiCalc}
\title{Fire Weather Index Calculation.}
\usage{
.fwiCalc(isi, bui)
}
\arguments{
\item{isi}{Initial Spread Index}

\item{bui}{Buildup Index}
}
\value{
A single fwi value
}
\description{
All code is based on a C code library that was written by Canadian
Forest Service Employees, which was originally based on
the Fortran code listed in the reference below. All equations
in this code refer to that document.
}
\details{
Equations and FORTRAN program for the Canadian Forest Fire 
Weather Index System. 1985. Van Wagner, C.E.; Pickett, T.L. 
Canadian Forestry Service, Petawawa National Forestry 
Institute, Chalk River, Ontario. Forestry Technical Report 33. 
18 p.

Additional reference on FWI system

Development and structure of the Canadian Forest Fire Weather 
Index System. 1987. Van Wagner, C.E. Canadian Forestry Service,
Headquarters, Ottawa. Forestry Technical Report 35. 35 p.
}
