% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fbp.r
\name{Fbp}
\alias{Fbp}
\title{Fire Behaviour Prediction System Calculation (hidden)}
\usage{
Fbp(input = NULL, output = "Primary")
}
\arguments{
\item{input}{Data frame of required and optional information needed to 
calculate FBP function. View the arguments section of the fbp manual (fbp.Rd) 
under "input" for the full listing of the required and optional inputs.}

\item{output}{What fbp outputs to return to the user. Options are "Primary", 
"Secondary" and "All". _Default:_ "Primary"}
}
\value{
output: Either Primary, Secondary, or all FBP outputs in a data.frame
}
\description{
Fire Behavior Prediction System calculations. This is the primary function 
for calculating FBP for a single timestep. Not all equations are calculated 
within this function, but have been broken down further.
}
