% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FineFuelMoistureCode.r
\name{FineFuelMoistureCode}
\alias{FineFuelMoistureCode}
\title{Fine Fuel Moisture Code Calculation}
\usage{
FineFuelMoistureCode(ffmc_yda, temp, rh, ws, prec)
}
\arguments{
\item{ffmc_yda}{The Fine Fuel Moisture Code from previous iteration}

\item{temp}{Temperature (centigrade)}

\item{rh}{Relative Humidity (%)}

\item{ws}{Wind speed (km/h)}

\item{prec}{Precipitation (mm)}
}
\value{
A single fine fuel moisture code value
}
\description{
Fine Fuel Moisture Code Calculation. All code is based on a C code library 
that was written by Canadian Forest Service Employees, which was originally 
based on the Fortran code listed in the reference below. All equations in 
this code refer to that document.
}
\details{
Equations and FORTRAN program for the Canadian Forest Fire Weather Index 
System. 1985. Van Wagner, C.E.; Pickett, T.L. Canadian Forestry Service, 
Petawawa National Forestry Institute, Chalk River, Ontario. Forestry 
Technical Report 33. 18 p.

Additional reference on FWI system Development and structure of the Canadian 
Forest Fire Weather Index System. 1987. Van Wagner, C.E. Canadian Forestry 
Service, Headquarters, Ottawa. Forestry Technical Report 35. 35 p.
}
