% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FireBehaviourPredictionC6.r
\name{FireBehaviourPredictionC6}
\alias{FireBehaviourPredictionC6}
\title{C-6 Conifer Plantaion Fire Spread Calculator}
\usage{
FireBehaviourPredictionC6(
  FUELTYPE,
  ISI,
  BUI,
  FMC,
  SFC,
  CBH,
  ROS,
  CFB,
  RSC,
  option = "CFB"
)
}
\arguments{
\item{FUELTYPE}{The Fire Behaviour Prediction FuelType}

\item{ISI}{Initial Spread Index}

\item{BUI}{Buildup Index}

\item{FMC}{Foliar Moisture Content}

\item{SFC}{Surface Fuel Consumption}

\item{CBH}{Crown Base Height}

\item{ROS}{Rate of Spread}

\item{CFB}{Crown Fraction Burned}

\item{RSC}{Crown Fire Spread Rate (m/min)}

\item{option}{Which variable to calculate(ROS, CFB, RSC, or RSI) _Default:_ "CFB"}
}
\value{
ROS, CFB, RSC or RSI depending on which option was selected
}
\description{
Calculate c6 (Conifer plantation) Fire Spread. C6 is a special case, and thus
 has it's own function. To calculate C6 fire spread, this function 
 also calculates and can return ROS, CFB, RSC, or RSI by specifying in 
 the option parameter.
All variables names are laid out in the same manner as Forestry Canada Fire 
Danger Group (FCFDG) (1992). Development and Structure of the Canadian Forest
 Fire Behavior Prediction System." Technical Report ST-X-3, Forestry Canada, 
 Ottawa, Ontario.
}
\references{
\url{https://cfs.nrcan.gc.ca/publications/download-pdf/10068} 
Development and Structure of the Canadian Forest Fire Behavior Prediction 
System." Technical Report ST-X-3, Forestry Canada, Ottawa, Ontario.
}
