% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlankRos.r
\name{FlankRos}
\alias{FlankRos}
\title{Flank Fire Rate of Spread Calculator}
\usage{
FlankRos(ROS, BROS, LB)
}
\arguments{
\item{ROS}{Fire Rate of Spread (m/min)}

\item{BROS}{Back Fire Rate of Spread (m/min)}

\item{LB}{Length to breadth ratio}
}
\value{
FROS Flank Fire Spread Rate (m/min) value
}
\description{
Calculate the Flank Fire Spread Rate.
}
\details{
All variables names are laid out in the same manner as Forestry Canada 
Fire Danger Group (FCFDG) (1992). Development and Structure of the 
Canadian Forest Fire Behavior Prediction System." Technical Report 
ST-X-3, Forestry Canada, Ottawa, Ontario.
}
