% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntermediateSurfaceRateOfSpreadC6.r
\name{IntermediateSurfaceRateOfSpreadC6}
\alias{IntermediateSurfaceRateOfSpreadC6}
\title{C-6 Conifer Plantation Intermediate Surface Fire Spread Rate Calculator}
\usage{
IntermediateSurfaceRateOfSpreadC6(ISI, FMC)
}
\arguments{
\item{ISI}{Initial Spread Index}

\item{FMC}{Foliar Moisture Content}
}
\value{
RSI
}
\description{
Calculate intermediate surface fire rate of spread, before BUI
effect is applied (RSI).
}
\details{
All variables names are laid out in the same manner as Forestry Canada Fire 
Danger Group (FCFDG) (1992). Development and Structure of the Canadian Forest
 Fire Behavior Prediction System." Technical Report ST-X-3, Forestry Canada, 
 Ottawa, Ontario.
}
\references{
\url{https://cfs.nrcan.gc.ca/publications/download-pdf/10068} 
Development and Structure of the Canadian Forest Fire Behavior Prediction 
System." Technical Report ST-X-3, Forestry Canada, Ottawa, Ontario.
}
