% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isi.r
\name{Isi}
\alias{Isi}
\title{Initial Spread Index Calculator}
\usage{
Isi(ffmc, ws, fbpMod = FALSE)
}
\arguments{
\item{ffmc}{Fine Fuel Moisture Code}

\item{ws}{Wind Speed (km/h)}

\item{fbpMod}{TRUE/FALSE if using the fbp modification at the extreme end}
}
\value{
ISI - Intial Spread Index
}
\description{
Computes the Initial Spread Index From the FWI System. Equations are from
Van Wagner (1985) as listed below, except for the modification for fbp
takene from FCFDG (1992).
}
\details{
Equations and FORTRAN program for the Canadian Forest Fire 
Weather Index System. 1985. Van Wagner, C.E.; Pickett, T.L. 
Canadian Forestry Service, Petawawa National Forestry 
Institute, Chalk River, Ontario. Forestry Technical Report 33. 
18 p.

Forestry Canada  Fire Danger Group (FCFDG) (1992). Development and 
Structure of the Canadian Forest Fire Behavior Prediction System."  
Technical ReportST-X-3, Forestry Canada, Ottawa, Ontario.
}
