% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lb.r
\name{Lb}
\alias{Lb}
\title{Length-to-Breadth ratio}
\usage{
Lb(FUELTYPE, WSV)
}
\value{
Length to Breadth ratio value
}
\description{
Computes the Length to Breadth ratio of an elliptically shaped fire. 
Equations are from listed FCFDG (1992) except for errata 80 from 
Wotton et. al. (2009).
}
\details{
All variables names are laid out in the same manner as Forestry Canada 
Fire Danger Group (FCFDG) (1992). Development and Structure of the 
Canadian Forest Fire Behavior Prediction System." Technical Report 
ST-X-3, Forestry Canada, Ottawa, Ontario.
  
Wotton, B.M., Alexander, M.E., Taylor, S.W. 2009. Updates and revisions to
the 1992 Canadian forest fire behavior prediction system. Nat. Resour. 
Can., Can. For. Serv., Great Lakes For. Cent., Sault Ste. Marie, Ontario, 
Canada. Information Report GLC-X-10, 45p.

 @param FUELTYPE The Fire Behaviour Prediction FuelType
 @param WSV The Wind Speed (km/h)
}
