% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LbT.r
\name{LbT}
\alias{LbT}
\title{Length-to-Breadth ratio at time t}
\usage{
LbT(FUELTYPE, LB, HR, CFB)
}
\arguments{
\item{FUELTYPE}{The Fire Behaviour Prediction FuelType}

\item{LB:}{Length to Breadth ratio}

\item{HR:}{Time since ignition (hours)}

\item{CFB:}{Crown Fraction Burned}
}
\value{
Length to Breadth ratio at time since ignition
}
\description{
Computes the Length to Breadth ratio of an elliptically shaped fire at
elapsed time since ignition. Equations are from listed FCFDG (1992) and
Wotton et. al. (2009), and are marked as such.
}
\details{
All variables names are laid out in the same manner as Forestry Canada 
Fire Danger Group (FCFDG) (1992). Development and Structure of the 
Canadian Forest Fire Behavior Prediction System." Technical Report 
ST-X-3, Forestry Canada, Ottawa, Ontario.

Wotton, B.M., Alexander, M.E., Taylor, S.W. 2009. Updates and revisions to
the 1992 Canadian forest fire behavior prediction system. Nat. Resour. 
Can., Can. For. Serv., Great Lakes For. Cent., Sault Ste. Marie, Ontario, 
Canada. Information Report GLC-X-10, 45p.
}
