% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateOfSpread.r
\name{RateOfSpread}
\alias{RateOfSpread}
\title{Rate of Spread Calculation}
\usage{
RateOfSpread(FUELTYPE, ISI, BUI, FMC, SFC, PC, PDF, CC, CBH)
}
\arguments{
\item{ISI}{Intiial Spread Index}

\item{BUI}{Buildup Index}

\item{FMC}{Foliar Moisture Content}

\item{SFC}{Surface Fuel Consumption (kg/m^2)}

\item{PC}{Percent Conifer (%)}

\item{PDF}{Percent Dead Balsam Fir (%)}

\item{CC}{Constant}

\item{CBH}{Crown to base height(m)}

\item{FUELTYPEThe}{Fire Behaviour Prediction FuelType}
}
\value{
ROS - Rate of Spread (m/min) value
}
\description{
Computes the Rate of Spread prediction based on fuel type and FWI
conditions. Equations are from listed FCFDG (1992) and Wotton et. al. 
(2009), and are marked as such.
}
\details{
All variables names are laid out in the same manner as Forestry Canada 
Fire Danger Group (FCFDG) (1992). Development and Structure of the 
Canadian Forest Fire Behavior Prediction System." Technical Report 
ST-X-3, Forestry Canada, Ottawa, Ontario.

Wotton, B.M., Alexander, M.E., Taylor, S.W. 2009. Updates and revisions to
the 1992 Canadian forest fire behavior prediction system. Nat. Resour. 
Can., Can. For. Serv., Great Lakes For. Cent., Sault Ste. Marie, Ontario, 
Canada. Information Report GLC-X-10, 45p.
}
