% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_RateOfSpreadAtTime.r
\name{RateOfSpreadAtTime}
\alias{RateOfSpreadAtTime}
\title{Rate of spread at time t calculation}
\usage{
RateOfSpreadAtTime(FUELTYPE, ROSeq, HR, CFB)
}
\arguments{
\item{FUELTYPE}{The Fire Behaviour Prediction FuelType}

\item{ROSeq}{Equilibrium Rate of Spread (m/min)}

\item{HR}{Time since ignition (hours)}

\item{CFB}{Crown Fraction Burned}
}
\value{
ROSt - Rate of Spread at time since ignition value
}
\description{
Computes the Rate of Spread prediction based on fuel type and FWI
conditions at elapsed time since ignition. Equations are from listed 
FCFDG (1992).
}
\details{
All variables names are laid out in the same manner as Forestry Canada 
Fire Danger Group (FCFDG) (1992). Development and Structure of the 
Canadian Forest Fire Behavior Prediction System." Technical Report 
ST-X-3, Forestry Canada, Ottawa, Ontario.
}
