% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateOfSpreadC6.r
\name{RateOfSpreadC6}
\alias{RateOfSpreadC6}
\title{C-6 Conifer Plantation Fire Spread Calculator}
\usage{
RateOfSpreadC6(RSC, RSS, CFB)
}
\arguments{
\item{RSC}{Crown Fire Spread Rate (m/min)}

\item{RSS}{Surface Fire Spread Rate (m/min)}

\item{CFB}{Crown Fraction Burned}
}
\value{
ROS
}
\description{
Calculate rate of spread (ROS).
}
\details{
All variables names are laid out in the same manner as Forestry Canada Fire 
Danger Group (FCFDG) (1992). Development and Structure of the Canadian Forest
 Fire Behavior Prediction System." Technical Report ST-X-3, Forestry Canada, 
 Ottawa, Ontario.
}
\references{
\url{https://cfs.nrcan.gc.ca/publications/download-pdf/10068} 
Development and Structure of the Canadian Forest Fire Behavior Prediction 
System." Technical Report ST-X-3, Forestry Canada, Ottawa, Ontario.
}
