% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_SlopeAdjust.r
\name{SlopeAdjust}
\alias{SlopeAdjust}
\title{Slope Adjusted wind speed or slope direction of spread calculation}
\usage{
SlopeAdjust(
  FUELTYPE,
  FFMC,
  BUI,
  WS,
  WAZ,
  GS,
  SAZ,
  FMC,
  SFC,
  PC,
  PDF,
  CC,
  CBH,
  ISI
)
}
\arguments{
\item{FUELTYPE}{The Fire Behaviour Prediction FuelType}

\item{BUI}{The Buildup Index value}

\item{WS}{Windspeed (km/h)}

\item{WAZ}{Wind Azimuth}

\item{GS}{Ground Slope (\%)}

\item{SAZ}{Slope Azimuth}

\item{FMC}{Foliar Moisture Content}

\item{SFC}{Surface Fuel Consumption (kg/m^2)}

\item{PC}{Percent Conifer (\%)}

\item{PDF}{Percent Dead Balsam Fir (\%)}

\item{CC}{Constant}

\item{CBH}{Crown Base Height (m)}

\item{ISI}{Initial Spread Index}

\item{output}{Type of variable to output (RAZ/WSV, default=RAZ)}
}
\value{
list(RAZ, WSV) - Rate of spread azimuth (degrees) and Wind Slope speed (km/hr)
}
\description{
Calculate the net effective windspeed (WSV), the net effective wind 
  direction (RAZ) or the wind azimuth (WAZ).
}
\details{
All variables names are laid out in the same manner as FCFDG (1992) and
  Wotton (2009).

  
  Forestry Canada Fire Danger Group (FCFDG) (1992). "Development and 
  Structure of the Canadian Forest Fire Behavior Prediction System." 
  Technical Report ST-X-3, Forestry Canada, Ottawa, Ontario.

  Wotton, B.M., Alexander, M.E., Taylor, S.W. 2009. Updates and revisions to
  the 1992 Canadian forest fire behavior prediction system. Nat. Resour. 
  Can., Can. For. Serv., Great Lakes For. Cent., Sault Ste. Marie, Ontario, 
  Canada. Information Report GLC-X-10, 45p.
}
