% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_SurfaceFuelConsumption.r
\name{SurfaceFuelConsumption}
\alias{SurfaceFuelConsumption}
\title{Surface Fuel Consumption Calculator}
\usage{
SurfaceFuelConsumption(FUELTYPE, FFMC, BUI, PC, GFL)
}
\arguments{
\item{FUELTYPE}{The Fire Behaviour Prediction FuelType}

\item{FFMC}{Fine Fuel Moisture Code}

\item{BUI}{Buildup Index}

\item{PC}{Percent Conifer (\%)}

\item{GFL}{Grass Fuel Load (kg/m^2)}
}
\value{
SFC Surface Fuel Consumption (kg/m^2)
}
\description{
Computes the Surface Fuel Consumption by Fuel Type.
  All variables names are laid out in the same manner as FCFDG (1992) or
  Wotton et. al (2009) 
  
  Forestry Canada Fire Danger Group (FCFDG) (1992). "Development and 
  Structure of the Canadian Forest Fire Behavior Prediction System." 
  Technical Report ST-X-3, Forestry Canada, Ottawa, Ontario.
}
\details{
Wotton, B.M., Alexander, M.E., Taylor, S.W. 2009. Updates and revisions to
  the 1992 Canadian forest fire behavior prediction system. Nat. Resour. 
  Can., Can. For. Serv., Great Lakes For. Cent., Sault Ste. Marie, Ontario, 
  Canada. Information Report GLC-X-10, 45p.
}
