% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TotalFuelConsumption.r
\name{TotalFuelConsumption}
\alias{TotalFuelConsumption}
\title{Total Fuel Consumption calculation}
\usage{
TotalFuelConsumption(FUELTYPE, CFL, CFB, SFC, PC, PDF, option = "TFC")
}
\arguments{
\item{FUELTYPE}{The Fire Behaviour Prediction FuelType}

\item{CFL}{Crown Fuel Load (kg/m^2)}

\item{CFB}{Crown Fraction Burned (0-1)}

\item{SFC}{Surface Fuel Consumption (kg/m^2)}

\item{PC}{Percent Conifer (%)}

\item{PDF}{Percent Dead Balsam Fir (%)}

\item{option}{Type of output (TFC, CFC, default=TFC)}
}
\value{
TFC Total (Surface + Crown) Fuel Consumption (kg/m^2) OR
CFC Crown Fuel Consumption (kg/m^2)
}
\description{
Computes the Total (Surface + Crown) Fuel Consumption by Fuel Type.
  All variables names are laid out in the same manner as FCFDG (1992) or
  Wotton et. al (2009) 
  
  Forestry Canada Fire Danger Group (FCFDG) (1992). "Development and 
  Structure of the Canadian Forest Fire Behavior Prediction System." 
  Technical Report ST-X-3, Forestry Canada, Ottawa, Ontario.
}
\details{
Wotton, B.M., Alexander, M.E., Taylor, S.W. 2009. Updates and revisions to
  the 1992 Canadian forest fire behavior prediction system. Nat. Resour. 
  Can., Can. For. Serv., Great Lakes For. Cent., Sault Ste. Marie, Ontario, 
  Canada. Information Report GLC-X-10, 45p.
}
